* commit 'afb3114606d96c247ad91cb2bd15dc7ed6e1579a':
  Switch the dimensions array to use uint32_t instead of size_t.

+++ cpu_ref/rsCpuIntrinsic.h
@@ -49,7 +49,7 @@ public:
 
     virtual void setGlobalVar(uint32_t slot, const void *data, size_t dataLength);
     virtual void setGlobalVarWithElemDims(uint32_t slot, const void *data, size_t dataLength,
-                                  const Element *e, const size_t *dims, size_t dimLength);
+                                  const Element *e, const uint32_t *dims, size_t dimLength);
     virtual void setGlobalBind(uint32_t slot, Allocation *data);
     virtual void setGlobalObj(uint32_t slot, ObjectBase *data);
